refactor: change Cycles and related clock fields from int to ulong
Promotes Cpu.Cycles, ClockEventEntry.Cycles, _nextEventCycle and all
peripheral-level cycle-tracking fields (Timer._lastCycle, EEPROM write
windows, Watchdog timeout, Clock._clockEnabledCycles) from int/uint to
ulong so the cycle counter never wraps after ~35 minutes at 16 MHz.
Arithmetic casts updated throughout Opcodes, peripherals and tests.

+++ src/Avr8Sharp.TestKit/Assertions/AvrCpuAssertions.cs
@@ -65,7 +65,7 @@ public class AvrCpuAssertions : ReferenceTypeAssertions<AvrCpu, AvrCpuAssertions
 
     /// <summary>Asserts that the cycle counter equals <paramref name="expected"/>.</summary>
     public AndConstraint<AvrCpuAssertions> HaveCycles(
-        int expected, string because = "", params object[] becauseArgs)
+        ulong expected, string because = "", params object[] becauseArgs)
     {
         Execute.Assertion
             .BecauseOf(because, becauseArgs)
