+++ tests/Avr8Sharp.Tests/CpuTests.cs
@@ -21,12 +21,12 @@ public class Cpu : AvrTestBase
 		[Test(Description = "The queued events should be executed after a given number of cycles")]
 		public void Execute_Queued_Events ()
 		{
-			var events = new List<KeyValuePair<int, int>> ();
+			var events = new List<KeyValuePair<int, ulong>> ();
 			int[] list = [1, 4, 10, ];
 			for (int i = 0; i < list.Length; i++) {
 				var value = list[i];
 				Cpu.AddClockEvent (() => {
-					events.Add (new KeyValuePair<int, int> (value, Cpu.Cycles));
+					events.Add (new KeyValuePair<int, ulong> (value, Cpu.Cycles));
 				}, value);
 			}
 			for (var i = 0; i < 10; i++) {
@@ -53,12 +53,12 @@ public class Cpu : AvrTestBase
 		[Test(Description = "The queued events should be correctly sorted when added in reverse order")]
 		public void Order_Reversed_Events ()
 		{
-			var events = new List<KeyValuePair<int, int>> ();
+			var events = new List<KeyValuePair<int, ulong>> ();
 			int[] list = [10, 4, 1, ];
 			for (int i = 0; i < list.Length; i++) {
 				var value = list[i];
 				Cpu.AddClockEvent (() => {
-					events.Add (new KeyValuePair<int, int> (value, Cpu.Cycles));
+					events.Add (new KeyValuePair<int, ulong> (value, Cpu.Cycles));
 				}, value);
 			}
 			for (var i = 0; i < 10; i++) {
@@ -90,13 +90,13 @@ public class Cpu : AvrTestBase
 			[Test(Description = "The cycles count should be updated according to the number of cycles of the given clock event")]
 			public void Update_Cycles_Count_Based_On_Clock_Event ()
 			{
-				var events = new List<KeyValuePair<int, int>> ();
+				var events = new List<KeyValuePair<int, ulong>> ();
 				var callbacks = new Dictionary<int, Action> ();
 				int[] list = [10, 4, 1, ];
 				for (int i = 0; i < list.Length; i++) {
 					var value = list[i];
 					callbacks[value] = Cpu.AddClockEvent (() => {
-						events.Add (new KeyValuePair<int, int> (value, Cpu.Cycles));
+						events.Add (new KeyValuePair<int, ulong> (value, Cpu.Cycles));
 					}, value);
 				}
 				Cpu.UpdateClockEvent (callbacks[4], 2);
@@ -130,13 +130,13 @@ public class Cpu : AvrTestBase
 				[Test(Description = "The clock event should be removed from the queue")]
 				public void Remove_Clock_Event ()
 				{
-					var events = new List<KeyValuePair<int, int>> ();
+					var events = new List<KeyValuePair<int, ulong>> ();
 					var callbacks = new Dictionary<int, Action> ();
 					int[] list = [1, 4, 10, ];
 					foreach (var value in list) {
 						var value1 = value;
 						callbacks[value] = Cpu.AddClockEvent (() => {
-							events.Add (new KeyValuePair<int, int> (value1, Cpu.Cycles));
+							events.Add (new KeyValuePair<int, ulong> (value1, Cpu.Cycles));
 						}, value);
 					}
 					Cpu.ClearClockEvent (callbacks[4]);
